requirements.txt
urllib3
json 
pandas
ast
pickle
whois
socket
websockets==10.4
uvicorn==0.20.0
verify-email==2.4.3
nest_asyncio
pysafebrowsing==0.1.2
absl-py==1.4.0
aiodns==3.0.0
aiosmtpd==1.4.4.post2
anyio==3.6.2
astunparse==1.6.3
atpublic==3.1.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
cachetools==5.3.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
cloud-init==22.4.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
configparser==5.3.0
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
ec2-hibinit-agent==1.0.0
fastapi==0.93.0
flatbuffers==23.3.3
future==0.18.3
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keras==2.11.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==15.0.6.1
mangum==0.17.0
Markdown==3.4.1
MarkupSafe==2.1.2
more-itertools==8.10.0
nest-asyncio==1.5.6
netifaces==0.11.0
numpy==1.24.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pexpect==4.8.0
protobuf==3.19.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycares==4.3.0
pycparser==2.21
pydantic==1.10.5
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pysafebrowsing==0.1.2
pyserial==3.5
python-apt==2.4.0
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-magic==0.4.24
python-whois==0.8.0
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.10.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sklearn==0.0.post1
sklearn-features==0.0.2
sniffio==1.3.0
sos==4.4
ssh-import-id==5.11
starlette==0.25.0
systemd-python==234
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
Twisted==22.1.0
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.20.0
verify-email==2.4.3
wadllib==1.3.6
Werkzeug==2.2.3
wrapt==1.15.0
zipp==1.0.0
zope.interface==5.4.0